+++ src/app/review/editReview/[reviewId]/page.tsx
@@ -1,0 +1,9 @@
+import Write from '@/components/Write/Write';
+
+export default function EditReview({ params }) {
+  return (
+    <>
+      <Write page={'editReview'} params={params} />
+    </>
+  );
+}
