+++ src/app/review/newReview/[reviewId]/page.tsx
@@ -1,0 +1,9 @@
+import Write from '@/components/Write/Write';
+
+export default function NewReview({ params }) {
+  return (
+    <>
+      <Write page={'newReview'} params={params} />
+    </>
+  );
+}
