+++ src/app/review/page.tsx
@@ -4,22 +4,25 @@ import Image from 'next/image';
 import RedWine from '../../../public/red-wine.svg';
 
 import Header from '@/components/Header/Header';
+import { useRouter } from 'next/navigation';
+
+const sampleData = [
+  {
+    id: '1',
+    date: '2024.00.00',
+    item_name: '토토 피에몬테 로쏘',
+    text: '와인이 너무 좋아요 즐겁고 빨개요',
+  },
+  {
+    id: '2',
+    date: '2024.01.01',
+    item_name: '토토 피에몬테 로쏘2',
+    text: '와인이 너무 좋아요 즐겁고 빨개요22',
+  },
+];
 
 const Review = () => {
-  const sampleData = [
-    {
-      id: '1',
-      date: '2024.00.00',
-      item_name: '토토 피에몬테 로쏘',
-      text: '와인이 너무 좋아요 즐겁고 빨개요',
-    },
-    {
-      id: '2',
-      date: '2024.01.01',
-      item_name: '토토 피에몬테 로쏘2',
-      text: '와인이 너무 좋아요 즐겁고 빨개요22',
-    },
-  ];
+  const route = useRouter();
 
   return (
     <>
@@ -31,7 +34,14 @@ const Review = () => {
               <div className="review_header_first_row">
                 <span className="review_date">{review.date}</span>
                 <div className="button_wrapper">
-                  <span className="button edit_button">수정</span>
+                  <span
+                    onClick={() => {
+                      route.push('review/editReview/1');
+                    }}
+                    className="button edit_button"
+                  >
+                    수정
+                  </span>
                   <span className="button delete_button">삭제</span>
                 </div>
               </div>
