+++ src/components/Write/Write.scss
@@ -1,0 +1,200 @@
+.WriteAsk {
+    min-height: 100vh;
+    width: 393px;
+    position: relative;
+  
+    &>.sendCheckBox {
+      position: fixed;
+      width: 393px;
+      height: 100%;
+      background: rgba(black, 0.5);
+      top: 0;
+      z-index: 9;
+      animation: opacityBg 0.5s;
+  
+      &>div {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        width: 100%;
+        height: 205px;
+        background: rgba(255, 255, 255, 1);
+        position: absolute;
+        bottom: 0px;
+        border-top-right-radius: 20px;
+        border-top-left-radius: 20px;
+  
+        &>h1 {
+          font-weight: 500;
+          font-size: 20px;
+          color: rgba(153, 153, 153, 1);
+          margin-bottom: 46px;
+        }
+  
+        &>button {
+          cursor: pointer;
+          width: 160px;
+          height: 45px;
+          font-weight: 700;
+          border: none;
+          background-color: rgba(0, 0, 0, 1);
+          color: rgba(255, 255, 255, 1);
+          border-radius: 30px;
+        }
+      }
+  
+      @keyframes opacityBg {
+        from {
+          opacity: 0%;
+        }
+  
+        to {
+          opacity: 100%;
+        }
+      }
+    }
+  }
+  
+  .WriteAsk-content {
+    padding: 0 20px;
+  
+    &>input {
+      width: 100%;
+      height: 42px;
+      border-radius: 20px;
+      border: 1px solid #d9d9d9;
+      font-size: 12px;
+      padding: 0 20px;
+      margin-bottom: 24px;
+  
+      &::placeholder {
+        color: rgba(153, 153, 153, 1);
+      }
+    }
+  
+    &>.textAreaBox {
+      position: relative;
+      font-size: 12px;
+      margin-bottom: 24px;
+  
+      &>textarea {
+        width: 100%;
+        min-height: 375px;
+        border-radius: 20px;
+        border: 1px solid rgba(217, 217, 217, 1);
+        padding: 18px;
+        overflow-y: scroll;
+        padding-bottom: 40px;
+        scroll-padding: 40px 0;
+        resize: none;
+        -ms-overflow-style: none;
+  
+        &::-webkit-scrollbar {
+          display: none;
+        }
+  
+        &::placeholder {
+          color: rgba(153, 153, 153, 1);
+        }
+      }
+  
+      &>p {
+        position: absolute;
+        bottom: 19px;
+        right: 19px;
+        color: rgba(153, 153, 153, 1);
+      }
+    }
+  
+    & .ask-images {
+      display: flex;
+      margin-bottom: 24px;
+  
+      &>img:not(:last-child) {
+        margin-right: 8px;
+      }
+    }
+  
+    & .add-image-button {
+      cursor: pointer;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+      margin-right: 8px;
+  
+      width: 63px;
+      height: 63px;
+      color: rgba(166, 166, 166, 1);
+      border: 1px solid rgba(217, 217, 217, 1);
+      font-size: 12px;
+      border-radius: 20px;
+  
+      &>p {
+        margin-top: 5px;
+      }
+  
+      &>.hidden {
+        display: none;
+      }
+    }
+  
+    &>.secret-writeBox {
+      color: rgba(153, 153, 153, 1);
+      font-size: 20px;
+  
+      &>input[type='checkbox'] {
+        display: none;
+      }
+  
+      &>input[type='checkbox']+label {
+        display: inline-block;
+        width: 16px;
+        height: 16px;
+        border-radius: 4px;
+        border: 1px solid rgba(217, 217, 217, 1);
+        margin-right: 13px;
+        position: relative;
+      }
+  
+      input[id='secret']:checked+label::after {
+        content: '✓';
+        width: 14px;
+        height: 10px;
+        text-align: center;
+        position: absolute;
+        color: rgba(255, 97, 53, 1);
+        left: -1px;
+        top: calc(50% - 12px);
+      }
+  
+      &>label {
+        cursor: pointer;
+      }
+    }
+  
+    &>.footer {
+      width: 100%;
+      height: 59px;
+      border-top: 1px solid #e3e3e3;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      position: absolute;
+      bottom: 0;
+      left: 0;
+      right: 0;
+  
+      &>button {
+        width: 345px;
+        height: 41px;
+        background-color: #ff6135;
+        font-weight: 700;
+        color: #ffffff;
+        font-size: 20px;
+        border-radius: 5px;
+        border: none;
+      }
+    }
+  }
